flickr: Factored out #write to module, now in _common/
The #write methods of old have been replaced with a module that
is included into classes that require writing. Then you declare what
can be written and some magic takes care of the rest.

In my opinion this is a lot nicer than the if-end waterfalls of
the previous implementation.

Also now experimenting with using hawx/clive to provide option
parsing.